refactor: 판매/구매하기 모달 따로 컴포넌트화
