+++ app/home/home-section.tsx
@@ -4,8 +4,7 @@ import React, { useState } from 'react';
 import { DataItemCard } from '../(shared)/components/DataItemCard';
 import { isToday } from '@/app/(shared)/utils/';
 import { PriceUnit } from '@/app/(shared)/types';
-import ModalPortal from '@/app/(shared)/components/modal-portal';
-import { useRouter } from 'next/navigation';
+import TradeConfirmationModal from '../(shared)/components/TradeConfirmationModal';
 
 interface DataItem {
   id: number;
@@ -50,7 +49,7 @@ const formatDataAmount = (amountInMB: number): string =>
 
 export default function HomeSection({ cards, unit }: HomeSectionProps) {
   const [modalItem, setModalItem] = useState<DataItem | null>(null);
-  const router = useRouter();
+
   return (
     <>
       {/* 카드 그리드 */}
@@ -72,70 +71,11 @@ export default function HomeSection({ cards, unit }: HomeSectionProps) {
         ))}
       </div>
 
-      <ModalPortal
-        isOpen={!!modalItem}
+      <TradeConfirmationModal
+        modalItem={modalItem}
         onClose={() => setModalItem(null)}
-        className="bg-black bg-opacity-50 flex items-center justify-center"
-      >
-        {modalItem && (
-          <div className="bg-white rounded-lg p-6 w-10/12 max-w-sm max-h-[60vh] overflow-y-auto">
-            <h3 className="text-regular-md font-bold text-left mb-1">
-              {modalItem.cardCategory === 'BUY' ? '판매하기' : '구매하기'}
-            </h3>
-
-            <p className="text-left text-regular-sm text-gray-600 mb-4">
-              아래 정보를 확인해주세요
-            </p>
-            <h2 className="text-heading-2xl font-bold text-center mb-2">
-              {formatCarrierName(modalItem.carrier)} 데이터{' '}
-              {formatDataAmount(modalItem.dataAmount)}{' '}
-            </h2>
-
-            <p className="text-center mb-6">
-              {modalItem.price.toLocaleString()}
-              {unit === 'snack' ? '스낵' : '₩'}에{' '}
-              {modalItem.cardCategory === 'BUY' ? '판매합니다' : '구매합니다'}.
-            </p>
-
-            <div className="mt-6 flex flex-col space-y-2">
-              <button
-                onClick={() => {
-                  if (modalItem.cardCategory === 'SELL') {
-                    const queryParams = new URLSearchParams({
-                      id: modalItem.id.toString(),
-                      pay: 'sell',
-                    });
-
-                    router.push(`/payment?${queryParams.toString()}`);
-                  } else {
-                    const confirmed = window.confirm('판매하시겠습니까?');
-                    if (confirmed) {
-                      console.log('판매 확정:', modalItem);
-                      // TODO: 구매자 로직 추가
-                    }
-                  }
-
-                  setModalItem(null);
-                }}
-                className={`w-full py-2 rounded ${
-                  modalItem.cardCategory === 'SELL'
-                    ? 'bg-midnight-black text-white'
-                    : 'bg-candy-pink text-white'
-                }`}
-              >
-                {modalItem.cardCategory === 'SELL' ? '구매하기' : '판매하기'}
-              </button>
-
-              <button
-                onClick={() => setModalItem(null)}
-                className="w-full py-2 rounded border"
-              >
-                취소하기
-              </button>
-            </div>
-          </div>
-        )}
-      </ModalPortal>
+        unit={unit}
+      />
     </>
   );
 }
